feat: [BE/FE] 투표 결과 기능 구현 (#25)

+++ storage/db-core/src/main/java/com/team/voteland/storage/db/core/vote/VoteOptionRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface VoteOptionRepository extends JpaRepository<VoteOptionEntity, Integer> {
+public interface VoteOptionRepository extends JpaRepository<VoteOptionEntity, Long> {
 
     long countByVoteId(Long voteId);
